Implements changes to adhere to MVVM pattern
